Moved some unused commands for ad into the dying core project
